Split home page into page/client
